pyproject.toml
[build-system]
build-backend = "poetry_dynamic_versioning.backend"
requires = [
  "poetry-core==1.6",
  "poetry-dynamic-versioning",
]

[tool.poetry]
name = "tap-readthedocs"
version = "0.0.0"
description = "`tap-readthedocs` is a Singer tap for ReadTheDocs, built with the Meltano SDK for Singer Taps."
authors = ["Edgar Ramírez-Mondragón <edgarrm358@gmail.com>"]
keywords = [
    "ELT",
    "ReadTheDocs",
    "singer.io",
]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/edgarrmondragon/tap-readthedocs"
repository = "https://github.com/edgarrmondragon/tap-readthedocs"
documentation = "https://github.com/edgarrmondragon/tap-readthedocs#readme"

[tool.poetry.dependencies]
python = "<3.12,>=3.7.1"
singer-sdk = "0.28.0"
toolz = "0.12.0"
requests-cache = "^1.0.1"

[tool.poetry.dev-dependencies]
singer-sdk = { version = "*", extras = ["testing"] }

[tool.poetry.scripts]
# CLI declaration
tap-readthedocs = 'tap_readthedocs.tap:TapReadTheDocs.cli'

[tool.pytest.ini_options]
addopts = "-vvv"

[tool.mypy]
python_version = 3.9
warn_unused_configs = true

[[tool.mypy.overrides]]
module = ["backoff.*", "toolz.*"]
ignore_missing_imports = true

[tool.poetry-dynamic-versioning]
enable = true
format-jinja = """
    {%- if distance == 0 -%}
        {{ serialize_pep440(base, stage, revision) }}
    {%- elif revision is not none -%}
        {{ serialize_pep440(base, stage, revision + 1, dev=distance, metadata=[commit]) }}
    {%- else -%}
        {{ serialize_pep440(bump_version(base), stage, revision, dev=distance, metadata=[commit]) }}
    {%- endif -%}
"""
metadata = true
style = "pep440"
vcs = "git"

[tool.ruff]
ignore = [
  "ANN101",  # missing-type-self
  "DJ",      # flake8-django
  "FIX002",  # line-contains-todo
]
line-length = 88
select = ["ALL"]
src = ["tap_pulumi_cloud", "tests"]
target-version = "py37"
unfixable = [
  "ERA001",  # commented-out-code
]

[tool.ruff.per-file-ignores]
"noxfile.py" = ["ANN"]
"tests/*" = [
  "ANN201",  # missing-return-type-public-function
  "S101",    # assert
  "SLF001",  # private-member-access
]

[tool.ruff.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.isort]
known-first-party = ["tap_readthedocs"]
required-imports = ["from __future__ import annotations"]

[tool.ruff.pydocstyle]
convention = "google"
